refactor: file을 관리하는 클래스 패키지 변경
