+++ src/main/java/file/AttendanceFileReader.java
@@ -1,4 +1,4 @@
-package domain;
+package file;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
