+++ src/test/java/AttendanceFileReaderTest.java
@@ -1,6 +1,6 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.AttendanceFileReader;
+import file.AttendanceFileReader;
 import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
